Refactor existing SR tests + test polishing
- test files hierarchy
- driver-specific fixtures
- multi-host tests separated into intrapool and crosspool tests
- single-host tests reworked to make better use of fixtures and remove
  the need for a specific execution order
- other minor changes
